Rename the include directory to use under_score naming, tweak CMakeLists
