+++ src/pid_controller_ros.cpp
@@ -1,4 +1,4 @@
-#include "PidController/pid_controller_ros.h"
+#include "pid_controller/pid_controller_ros.h"
 
 PidControllerRos::PidControllerRos()
                 : reference_(0.0),
